+++ VadereSimulator/src/org/vadere/simulator/projects/migration/incident/IncidentDatabase.java
@@ -7,7 +7,7 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 
-import org.vadere.simulator.entrypoints.Version;
+import org.vadere.util.version.Version;
 import org.vadere.simulator.projects.migration.incident.incidents.AddTextNodeIncident;
 import org.vadere.simulator.projects.migration.incident.incidents.DeletionIncident;
 import org.vadere.simulator.projects.migration.incident.incidents.Incident;
@@ -20,7 +20,7 @@ import org.vadere.simulator.projects.migration.incident.incidents.specialized.Mo
 import org.vadere.state.attributes.AttributesSimulation;
 import org.vadere.state.util.StateJsonConverter;
 
-import static org.vadere.simulator.entrypoints.Version.*;
+import static org.vadere.util.version.Version.*;
 
 public class IncidentDatabase {