+++ VadereSimulator/src/org/vadere/simulator/projects/migration/jsontranformation/jolt/JoltTransformV4toV5.java
@@ -3,9 +3,8 @@ package org.vadere.simulator.projects.migration.jsontranformation.jolt;
 import com.fasterxml.jackson.databind.JsonNode;
 
 import org.vadere.annotation.factories.migrationassistant.MigrationTransformation;
-import org.vadere.simulator.entrypoints.Version;
+import org.vadere.util.version.Version;
 import org.vadere.simulator.projects.migration.MigrationException;
-import org.vadere.simulator.projects.migration.jsontranformation.AbstractJsonTransformation;
 import org.vadere.simulator.projects.migration.jsontranformation.JoltTransformation;
 
 import java.util.Iterator;