+++ VadereSimulator/src/org/vadere/simulator/projects/migration/jsontranformation/json/TargetVersionV1_6.java
@@ -2,14 +2,10 @@ package org.vadere.simulator.projects.migration.jsontranformation.json;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
-import org.apache.commons.lang3.tuple.Pair;
 import org.vadere.annotation.factories.migrationassistant.MigrationTransformation;
-import org.vadere.simulator.entrypoints.Version;
+import org.vadere.util.version.Version;
 import org.vadere.simulator.projects.migration.MigrationException;
 import org.vadere.simulator.projects.migration.jsontranformation.SimpleJsonTransformation;
-
-import java.util.ArrayList;
-import java.util.List;
 
 /**
  * Rename following pedestrian attributes: