+++ VadereSimulator/tests/org/vadere/simulator/entrypoints/MigrationSubCommandTest.java
@@ -11,6 +11,7 @@ import org.vadere.simulator.projects.migration.MigrationAssistant;
 import org.vadere.simulator.utils.reflection.TestJsonNodeExplorer;
 import org.vadere.simulator.utils.reflection.TestJsonNodeHelper;
 import org.vadere.simulator.utils.reflection.TestResourceHandlerScenario;
+import org.vadere.util.version.Version;
 
 import java.nio.file.Files;
 import java.nio.file.Path;