+++ VadereSimulator/tests/org/vadere/simulator/projects/migration/jsontranformation/JoltTransformV1toV2Test.java
@@ -3,7 +3,7 @@ package org.vadere.simulator.projects.migration.jsontranformation;
 import com.fasterxml.jackson.databind.JsonNode;
 
 import org.junit.Test;
-import org.vadere.simulator.entrypoints.Version;
+import org.vadere.util.version.Version;
 import org.vadere.simulator.projects.migration.MigrationException;
 
 import java.nio.file.Path;