+++ VadereSimulator/tests/org/vadere/simulator/projects/migration/jsontranformation/json/V07ToV08Test.java
@@ -3,7 +3,7 @@ package org.vadere.simulator.projects.migration.jsontranformation.json;
 import com.fasterxml.jackson.databind.JsonNode;
 
 import org.junit.Test;
-import org.vadere.simulator.entrypoints.Version;
+import org.vadere.util.version.Version;
 import org.vadere.simulator.projects.migration.MigrationException;
 import org.vadere.simulator.projects.migration.jsontranformation.JsonTransformationTest;