+++ VadereSimulator/tests/org/vadere/simulator/utils/reflection/TestJsonNodeExplorer.java
@@ -6,7 +6,7 @@ import org.hamcrest.BaseMatcher;
 import org.hamcrest.Description;
 import org.hamcrest.Matcher;
 import org.hamcrest.core.StringContains;
-import org.vadere.simulator.entrypoints.Version;
+import org.vadere.util.version.Version;
 import org.vadere.simulator.projects.migration.jsontranformation.JsonNodeExplorer;
 
 import java.util.ArrayList;